+++ packages/tswebextension/test/lib/mv2/background/injection-helper.test.ts
@@ -7,7 +7,7 @@ jest.spyOn(Date, 'now').mockReturnValueOnce(timestamp);
 const trim = (str: string): string => str.replace(/\s+/g, '');
 
 const expectedScriptTemplate = (text: string): string => `(function() {\
-    if (window.scriptExecuted${timestamp}) {\
+    if (window.scriptExecuted${timestamp} || document instanceof XMLDocument) {\
         return;\
     }\
     var script = document.createElement("script");\