+++ packages/adguard-api/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.2] - 2024-02-08
+
+### Added
+- New private method `removeObsoletedFilters` to prevent usage of obsoleted filters.
+- New public event channel `AdguardApi.onFiltersDeletion`.
+
+### Changed
+- `adguardApi.start` returns Promise with applied configuration with already removed
+obsoleted filters ids.
+
 ## [2.1.1] - 2023-12-01
 
 ### Added